Binary Search
This program is designed for searching an item in an array using binary search. Initially, it prompts the user for the size of the array and its elements. Subsequently, it employs the selection sort algorithm to arrange the array. The program then requests the user to input the item to be searched. If the item is found, it displays a message indicating the item's index number. Otherwise, it prints a message stating that the item was not found.