Bubble Sort Algorithm
This program initially prompts the user to input the size of the array. If the size is less than '0,' it asks repeatedly until the user inputs a positive integer. It then asks for array elements, sorts the array through the Bubble Sort algorithm, and displays all elements of the array.