Deleting element from Array
This program prompts the user to enter the size of the array. It then asks the user to enter array elements. After that, it asks for the position of the element to be deleted. The position is then checked. If the position is greater or smaller than the size of the array, it prints an invalid message. Once the user inputs a valid position, the program deletes the element at that position and displays the new array.